FilterLocalization now takes Entities instead of dtos
